Added support for custom unit definitions.
This requires adopting a pint-like approach where the UnitRegistry
provides special Unit and Quantity classes that contain references
to the parent UnitRegistry.

To keep this performant, unpickling Quantity now injects a globally
set UnitRegistry instance, which will usually be the default set of
unit definitions. This can be overwritten by the user with
ApplicationRegistry.set.